Manage the data obj on the separated package
